Corrección int>float

+++ Guia2ej6.cpp
@@ -5,7 +5,7 @@ using namespace std;
 // Ejercicio 6: Escriba un programa que calcule x^y, donde tanto x como y son enteros positivos, sin utilizar la función pow.
 
 int main() {
-    int x, y, pot = 1;
+    float x, y, pot = 1;
     cout <<"Ingrese primero un valor inicial y despues a que numero lo quiere elevar (ambos deben ser numero enteros positivos)"<< endl;
     cin >> x;
     cin >> y;
